- Use ClientProxyBuilder instead of custom code - Refactored config package so it is in the project root as per DANS CST conventions.
