+++ src/main/java/nl/knaw/dans/validatedansbag/DdValidateDansBagApplication.java
@@ -22,7 +22,7 @@ import io.dropwizard.core.setup.Environment;
 import io.dropwizard.forms.MultiPartBundle;
 import nl.knaw.dans.lib.util.ClientProxyBuilder;
 import nl.knaw.dans.validatedansbag.client.VaultCatalogClientImpl;
-import nl.knaw.dans.validatedansbag.core.config.DdValidateDansBagConfiguration;
+import nl.knaw.dans.validatedansbag.config.DdValidateDansBagConfiguration;
 import nl.knaw.dans.validatedansbag.core.engine.RuleEngineImpl;
 import nl.knaw.dans.validatedansbag.core.rules.RuleSets;
 import nl.knaw.dans.validatedansbag.core.service.BagItMetadataReaderImpl;
