+++ src/main/java/nl/knaw/dans/validatedansbag/config/XmlSchemaConfig.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.validatedansbag.core.config;
+package nl.knaw.dans.validatedansbag.config;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Getter;
